Factor the signup with password view
