+++ src/widgets/auth/authWidget.js
@@ -6,6 +6,7 @@ import LoginView from './views/loginViewComponent'
 import LoginWithWebAuthnView from './views/loginWithWebAuthnViewComponent'
 import LoginWithPasswordView from './views/loginWithPasswordViewComponent'
 import SignupView from './views/signupViewComponent'
+import SignupWithPasswordView from './views/signupWithPasswordViewComponent'
 import { ForgotPasswordView, ForgotPasswordSuccessView } from './views/forgotPasswordViewComponent'
 import QuickLoginView from './views/quickLoginViewComponent'
 import ReauthView from './views/reauthViewComponent'
@@ -30,6 +31,7 @@ export default createMultiViewWidget({
         'login-with-web-authn': LoginWithWebAuthnView,
         'login-with-password': LoginWithPasswordView,
         'signup': SignupView,
+        'signup-with-password': SignupWithPasswordView,
         'forgot-password': ForgotPasswordView,
         'forgot-password-success': ForgotPasswordSuccessView,
         'quick-login': QuickLoginView,
